wip
ok lets go with serilog and don't do crap like this

        // ... rest of the methods remain the same ...

I literally just copy pasted the whole thing and I see this?

also please do this for both program.cs as well as for the library and please implement the versioning in the directory.something.props file so versions go in the same central place.

also please clean up for solid principles and best practices for example class library should always ilogger and never console.writeline. in fact, now that we have serilog, I don't see why even the console should console.writeline. I think we still need console.readline but all console.writeline can be removed and we can write everything to serilog and have serilog save to daily log files in text files as well as on the screen?